Some polishing and one more input constructor
